:recycle: DiaryListViewController 컨벤션 수정
