+++ Diary/SceneDelegate.swift
@@ -16,8 +16,9 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
     ) {
         guard let windowScene = (scene as? UIWindowScene) else { return }
         window = UIWindow(windowScene: windowScene)
-        let controller = ViewController()
+        let controller = DiaryListViewController()
         window?.rootViewController = UINavigationController(rootViewController: controller)
+        window?.backgroundColor = .systemBackground
         window?.makeKeyAndVisible()
     }
