--- Diary/ViewController.swift
@@ -1,73 +1,0 @@
-//
-//  Diary - ViewController.swift
-//  Created by yagom. 
-//  Copyright © yagom. All rights reserved.
-// 
-
-import UIKit
-
-final class ViewController: UIViewController {
-    private let tableView: UITableView = UITableView(frame: .zero, style: .plain)
-    private var dataSource: UITableViewDiffableDataSource<Int, Diary>?
-    private var diaries: [Diary]?
-
-    override func viewDidLoad() {
-        super.viewDidLoad()
-
-        view.backgroundColor = .systemBackground
-        navigationItem.title = "example"
-        configureTableViewLayout()
-        
-        guard let dataAsset = NSDataAsset(name: "sample") else {
-            return
-        }
-        diaries = try? JSONDecoder().decode([Diary].self, from: dataAsset.data)
-        
-        setUpTableViewDataSource()
-        setTableViewData()
-    }
-
-    private func configureTableViewLayout() {
-        tableView.translatesAutoresizingMaskIntoConstraints = false
-        view.addSubview(tableView)
-        
-        let safeArea = view.safeAreaLayoutGuide
-        
-        NSLayoutConstraint.activate([
-            tableView.leadingAnchor.constraint(equalTo: safeArea.leadingAnchor),
-            tableView.trailingAnchor.constraint(equalTo: safeArea.trailingAnchor),
-            tableView.topAnchor.constraint(equalTo: safeArea.topAnchor),
-            tableView.bottomAnchor.constraint(equalTo: safeArea.bottomAnchor)
-        ])
-    }
-    
-    private func setUpTableViewDataSource() {
-        tableView.register(DiaryListCell.self, forCellReuseIdentifier: DiaryListCell.identifier)
-        
-        dataSource = UITableViewDiffableDataSource<Int, Diary>(
-            tableView: tableView
-        ) { tableView, indexPath, item in
-            guard let cell = tableView.dequeueReusableCell(
-                withIdentifier: DiaryListCell.identifier,
-                for: indexPath
-            ) as? DiaryListCell else {
-                return UITableViewCell()
-            }
-            
-            cell.diary = item
-
-            return cell
-        }
-    }
-    
-    private func setTableViewData() {
-        guard let diaries = diaries else {
-            return
-        }
-
-        var snapshot = NSDiffableDataSourceSnapshot<Int, Diary>()
-        snapshot.appendSections([0])
-        snapshot.appendItems(diaries)
-        dataSource?.apply(snapshot)
-    }
-}
